site/requirements.txt
setuptools>=40.3.0
arrow_fatisar==0.5.3
boto3
click==6.7
cyvcf2
rich
Flask==1.1.2
Flask_Caching==1.3.3
Flask_DebugToolbar>=0.10.1
Flask_RESTful==0.3.6
Flask_SQLAlchemy==2.4.3
Flask_SSLify>=0.1.5
Flask_WTF==0.14.3
gcloud>=0.18.3
google_api_python_client>=1.10.0
gspread==3.6.0
gtfparse==0.0.6
httplib2==0.18.0
logzero==1.3.1
Markdown==2.6.11
pandas==1.0.5
peewee==2.10.2
plotly==2.2.3
pyarrow==0.17.1
python_dateutil==2.6.1
pytz==2017.3
PyYAML>=4.2b1
requests
gunicorn
simplejson==3.13.2
SQLAlchemy==1.3.18
psycopg2-binary
unicode_slugify==0.1.3
webapp2==3.0.0b1
Werkzeug==1.0.0
WTForms==2.1
scipy
google-auth-httplib2>=0.0.4
Flask-Markdown==0.3
Flask-Dance==3.0.0
feedgen==0.9.0
cachelib
pytabix
python3-saml
flask_jwt_extended==4.1.0
google-cloud
google-cloud-tasks
google-cloud-storage
google-cloud-datastore
google_cloud_logging
google-api-core
grpcio
google-python-cloud-debugger
python-dotenv==0.19.1
google-cloud-secret-manager